+++ src/TransmissionExtras.Server/Program.cs
@@ -34,12 +34,7 @@ builder.Services.AddHostedService<VerifyTorrentsJob>();
 
 var app = builder.Build();
 
-var summaries = new[]
-{
-    "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-};
-
-app.MapGet("api/healthcheck", async Task<Results<Ok, ProblemHttpResult>> ([FromServices] IOptions<TransmissionOptions> options) =>
+app.MapGet("api/healthcheck", async Task<Results<Ok<HealthCheckResult>, ProblemHttpResult>> ([FromServices] IOptions<TransmissionOptions> options) =>
 {
     try
     {
@@ -50,7 +45,7 @@ app.MapGet("api/healthcheck", async Task<Results<Ok, ProblemHttpResult>> ([FromS
 
         _ = await client.GetSessionInformationAsync();
 
-        return TypedResults.Ok();
+        return TypedResults.Ok<HealthCheckResult>(new("Ok"));
     }
     catch (Exception ex)
     {
@@ -58,19 +53,6 @@ app.MapGet("api/healthcheck", async Task<Results<Ok, ProblemHttpResult>> ([FromS
 
         return TypedResults.Problem(title: "Healthchecks failed", detail: ex.Message);
     }
-});
-
-app.MapGet("/weatherforecast", () =>
-{
-    var forecast = Enumerable.Range(1, 5).Select(index =>
-        new WeatherForecast
-        (
-            DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
-            Random.Shared.Next(-20, 55),
-            summaries[Random.Shared.Next(summaries.Length)]
-        ))
-        .ToArray();
-    return forecast;
 });
 
 try
@@ -100,11 +82,8 @@ partial class Program
     static partial void LogHealthCheckFailed(ILogger logger, Exception e);
 }
 
-internal record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
-{
-    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-}
+record HealthCheckResult(string Status);
 
-[JsonSerializable(typeof(WeatherForecast[]))]
+[JsonSerializable(typeof(HealthCheckResult))]
 
 internal partial class AppJsonSerializerContext : JsonSerializerContext { }